kFile Explorer, Funktionen alle ausgelagert
